+++ packages/ds/react/lib/stencil-generated/components.ts
@@ -1,0 +1,11 @@
+/* eslint-disable */
+/* tslint:disable */
+/* auto-generated react proxies */
+import { createReactComponent } from './react-component-lib';
+
+import type { JSX } from '@ircc-ca/ds-sdc-web';
+
+
+
+export const DjlButton = /*@__PURE__*/createReactComponent<JSX.DjlButton, HTMLDjlButtonElement>('djl-button');
+export const MyComponent = /*@__PURE__*/createReactComponent<JSX.MyComponent, HTMLMyComponentElement>('my-component');