+++ packages/ds/web/stencil.config.ts
@@ -1,5 +1,6 @@
 import { angularOutputTarget } from '@stencil/angular-output-target';
 import { Config } from '@stencil/core';
+import { reactOutputTarget } from '@stencil/react-output-target';
 import { sass } from '@stencil/sass';
 
 export const config: Config = {
@@ -23,7 +24,11 @@ export const config: Config = {
     angularOutputTarget({
       componentCorePackage: '@ircc-ca/ds-sdc-web',
       directivesProxyFile: '../angular/src/stencil-generated/components.ts',
-      directivesArrayFile: '../angular/src/stencil-generated/index.ts',
+    }),
+    reactOutputTarget({
+      componentCorePackage: '@ircc-ca/ds-sdc-web',
+      proxiesFile: '../react/lib/stencil-generated/components.ts',
+      loaderDir: 'dist/loader',
     }),
   ],
 };