+++ stories/Button.stories.ts
@@ -1,14 +1,64 @@
 import { CommonModule } from '@angular/common';
-import { ButtonComponent } from '@ircc-ca/ds-sdc-angular/button';
-import { IconComponent, IIconConfig } from '@ircc-ca/ds-sdc-angular/icon';
+import { IRCCModule } from '@ircc-ca/ds-sdc-angular';
 import { moduleMetadata } from '@storybook/angular';
 import { Meta, Story } from '@storybook/angular/types-6-0';
 
+enum Color {
+    critical = 'critical',
+    cta = 'CTA',
+}
+enum Type {
+    button = 'button',
+    submit = 'submit',
+    reset = 'reset',
+}
+enum Category {
+    primary = 'primary',
+    secondary = 'secondary',
+    plain = 'plain',
+}
+
+enum Size {
+    small = 'small',
+    large = 'large',
+}
+interface ButtonProps {
+    disabled?: boolean;
+    name?: string;
+    color?: Color | string;
+    type?: Type | string;
+    value?: string;
+    category?: Category | string;
+    size?: Size | string;
+    ariaLabel?: string;
+}
+
 export default {
     title: 'Components/Button',
-    component: ButtonComponent,
     args: {
         ngContent: 'Label',
+    },
+    argTypes: {
+        color: {
+            description: 'Colors for button (optional)',
+            options: [Color.cta, Color.critical],
+            control: { type: 'radio' },
+        },
+        size: {
+            description: 'Size for button (optional)',
+            options: [Size.small, Size.large],
+            control: { type: 'radio' },
+        },
+        category: {
+            description: 'Category for button (optional)',
+            options: [Category.primary, Category.secondary, Category.plain],
+            control: { type: 'radio' },
+        },
+        type: {
+            description: 'Type for button (optional)',
+            options: [Type.button, Type.reset, Type.submit],
+            control: { type: 'radio' },
+        },
     },
     parameters: {
         design: {
@@ -18,97 +68,34 @@ export default {
     },
     decorators: [
         moduleMetadata({
-            declarations: [ButtonComponent, IconComponent],
-            imports: [CommonModule],
+            imports: [CommonModule, IRCCModule],
         }),
     ],
-} as Meta<ButtonComponent>;
+} as Meta;
 
-const Template: Story<ButtonComponent> = (args) => ({
+const Template: Story<ButtonProps> = (args: ButtonProps) => ({
     props: { ...args },
-    template: `<jds-button 
-    [category]="category"
-    [ariaLabel]="ariaLabel"
-    [color]="color"
-    [size]="size"
-    [type]="type"
-    [disabled]="disabled"
-    >{{ ngContent }}</jds-button>`,
+    template: `<djl-button
+        color="${args.color}"
+        category="${args.category}"
+        size="${args.size}"
+        disabled="${args.disabled}"
+        type="${args.type}"
+        aria-label="${args.ariaLabel}"
+        name="${args.name}"
+        value="${args.value}"
+        >{{ngContent}}</djl-button
+    >`,
 });
 
-export const Primary = Template.bind({});
-Primary.args = {
-    ariaLabel: 'aria text',
-    category: 'primary',
+export const Default = Template.bind({});
+Default.args = {
+    color: Color.cta,
+    category: Category.primary,
     disabled: false,
-};
-
-export const Secondary = Template.bind({});
-Secondary.args = {
-    ariaLabel: 'aria text',
-    category: 'secondary',
-    disabled: false,
-};
-
-export const Plain = Template.bind({});
-Plain.args = {
-    ariaLabel: 'aria text',
-    category: 'plain',
-    disabled: false,
-};
-
-export const Disabled = Template.bind({});
-Disabled.args = {
-    category: 'primary',
-    disabled: true,
-};
-
-export const ButtonWithIconAndText: Story<IconComponent> = (args) => ({
-    props: {
-        ...args,
-    },
-    template: `<jds-button 
-  [category]="category"
-  [ariaLabel]="ariaLabel"
-  [color]="color"
-  [size]="size"
-  [type]="type"
-  [disabled]="disabled"
-  >
-  <jds-font-icon [iconConfig]="iconConfig"></jds-font-icon>
-  {{ ngContent }}</jds-button>`,
-});
-
-const BtnConfig: IIconConfig = {
-    fontFamily: 'fa-solid',
-    unicode: 'f6b0',
-};
-
-ButtonWithIconAndText.args = {
-    iconConfig: BtnConfig,
-};
-
-export const ButtonWithIcon: Story<IconComponent> = (args) => ({
-    props: {
-        ...args,
-    },
-    template: `<jds-button 
-  [category]="category"
-  [ariaLabel]="ariaLabel"
-  [color]="color"
-  [size]="size"
-  [type]="type"
-  [disabled]="disabled"
-  >
-  <jds-font-icon [iconConfig]="iconConfig"></jds-font-icon>
-  </jds-button>`,
-});
-
-const BtnIconConfig: IIconConfig = {
-    fontFamily: 'fa-solid',
-    unicode: 'f6b0',
-};
-
-ButtonWithIcon.args = {
-    iconConfig: BtnIconConfig,
+    size: Size.small,
+    type: Type.button,
+    ariaLabel: 'This is an aria-label',
+    name: 'Stencil btn',
+    value: 'stencil',
 };